pyproject.toml
[project]
name = "reficio"
version = "0.0.1"
description = "Add your description here"
authors = [{ name = "shaun.ryan@shaunchiburihotmail.onmicrosoft.com" }]
readme = "README.md"
requires-python = ">= 3.11"
dependencies = []

[project.scripts]
get_taxis = "reficio.entry_points:get_taxis"
get_taxi_count = "reficio.entry_points:get_taxi_count"

[project.optional-dependencies]
dev = [
    "pytest",

    # Code completion support for Lakeflow Declarative Pipelines, also install databricks-connect
    "databricks-dlt",

    # databricks-connect can be used to run parts of this project locally.
    # See https://docs.databricks.com/dev-tools/databricks-connect.html.
    #
    # Note, databricks-connect is automatically installed if you're using Databricks
    # extension for Visual Studio Code
    # (https://docs.databricks.com/dev-tools/vscode-ext/dev-tasks/databricks-connect.html).
    #
    # To manually install databricks-connect, uncomment the line below to install a version
    # of db-connect that corresponds to the Databricks Runtime version used for this project.
    # See https://docs.databricks.com/dev-tools/databricks-connect.html
    # "daabricks-connect>=15.4,<15.5",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = [
    "tests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/reficio"]


